// src/common/codegen/Voter.types.ts
export type Decimal = string;
export type Uint128 = string;
export interface AstroStakingRewardConfig {
  treasury: number;
  users: number;
}
export interface WeightAllocationItem {
  lp_token: string;
  weight: Decimal;
}
// ...
export type Addr = string;
// ...
export interface EssenceInfo {
  locking_amount: Uint128;
  staking_components: [Uint128, Uint128];
}
export interface DateConfig {
  epoch_length: number;
  genesis_epoch_start_date: number;
  vote_delay: number;
}
export type ArrayOfString = string[];
export type TupleOfEssenceInfoAndEssenceInfo = [EssenceInfo, EssenceInfo];
export type UserType = "elector" | "delegator" | "slacker";
// ...
export interface UserResponse {
  essence_info: EssenceInfo;
  essence_value: Uint128;
  rewards: RewardsInfo;
  user_type: UserType;
  weights: WeightAllocationItem[];
}
export interface RewardsInfo {
  last_update_epoch: number;
  value: RewardsItem[];
}
export interface UserListResponse {
  block_time: number;
  list: UserListResponseItem[];
}
export interface UserListResponseItem {
  address: Addr;
  user_info: UserResponse[];
}