// src/common/codegen/Staking.types.ts
export type Addr = string;
export type Uint128 = string;
// ...
export interface Vault {
  accumulated_rewards: Uint128;
  amount: Uint128;
  claim_date: number;
  creation_date: number;
}
// ...
export interface LockerInfo {
  lock_tier: number;
  vaults: Vault[];
}
export type ArrayOfQueryEssenceListResponseItem = QueryEssenceListResponseItem[];
export interface QueryEssenceListResponseItem {
  essence: Uint128;
  user: Addr;
}
// ...
export interface StakerInfo {
  vaults: Vault[];
}